Split simlulator params to basic/advanced
