+++ apps/site-landing/src/containers/BenefitsSimulator/Form/index.tsx
@@ -1,69 +1,16 @@
-import React, { Dispatch, SetStateAction } from 'react';
-import { Button } from 'semantic-ui-react';
-import { defineMessages, FormattedMessage } from 'react-intl';
-import { SimulationParameters } from '../../ReturnProfile/data/params';
-import { RangeFieldAndScale } from 'components/RangeFieldAndScale';
+import React from 'react';
+import { Tab } from 'semantic-ui-react';
+import { Props } from './types';
+import { Basic } from './Basic';
+import { Advanced } from './Advanced';
 import styles from './styles.module.less';
-import { debounce } from 'lodash';
 
-const translations = defineMessages({
-    startingValue : {
-      defaultMessage: 'Starting value:',
-      description: 'site.compare.label.rangeStarting: label for Starting Value in the compare page'},
-    rangeDuration : {
-      defaultMessage: 'Duration:',
-      description: 'site.compare.label.rangeDuration: label for range duration in the compare page'},
-    rangeDurationYear : {
-      defaultMessage: 'Years',
-      description: 'site.compare.label.rangeDurationYear: label for range Duration Year field in the compare page'},
-    button : {
-        defaultMessage: 'Show Me',
-        description: 'site.compare.button: label for the button in the compare page'}
-  });
+export const Form = (props: Props) => {
 
-type Props = {
-  params: SimulationParameters;
-  setParams: Dispatch<SetStateAction<SimulationParameters>>;
-  years: number,
-  setYears: Dispatch<SetStateAction<number>>;
-};
+  const panes = [
+    { menuItem: 'Basic', render: () => <Basic {...props} /> },
+    { menuItem: 'Advanced', render: () => <Advanced {...props} /> },
+  ]
 
-export const FormCompare = ({params, setParams, years, setYears}: Props) => {
-
-  const debounceInterval = 250;
-  const dbOnChangeNamed = (name: keyof SimulationParameters) =>
-    debounce((val: any) => setParams((prev: SimulationParameters) => ({
-      ...prev,
-      [name]: val,
-    })), debounceInterval);
-
-  const dbOnSetYears = debounce(v => setYears(Math.max(1, v)), debounceInterval);
-
-  return (
-    <div id={styles.variablesContainer}>
-
-      <RangeFieldAndScale
-        label={translations.startingValue}
-        scaleType="currency"
-        currency="CAD"
-        maximum={1000}
-        step={1}
-        initial={params.initialBalance}
-        onChange={dbOnChangeNamed("initialBalance")}
-      />
-
-      <RangeFieldAndScale
-        label={translations.rangeDuration}
-        scaleType="unit"
-        unit="year"
-        maximum={60}
-        initial={years}
-        onChange={dbOnSetYears}
-      />
-
-      <Button secondary className={`${styles.buttonContainer} x16spaceBefore`}>
-        <FormattedMessage {...translations.button} />
-      </Button>
-    </div>
-  );
+  return <Tab className={styles.container} panes={panes} />;
 };
