+++ apps/site-landing/src/containers/BenefitsSimulator/Form/styles.module.less
@@ -1,14 +1,22 @@
-#variablesContainer{
-    background: url("images/background.svg");
-    background-position: right bottom;
-    max-width: 370px;
-    height: 550px;
-    background-repeat: none;
-    border-radius: 5px;
-    color: white;
-  
-    .buttonContainer{
-      bottom: 10px;
-      position:relative;
-    }
+.container {
+  max-width: 370px;
+  height: 550px;
+}
+
+.formPane {
+  height: 100%;
+  background: url("images/background.svg");
+  background-position: right bottom;
+
+  padding: 1rem;
+  background-repeat: none;
+  border-bottom-left-radius: 5px;
+  border-bottom-right-radius: 5px;
+  border-top-right-radius: 5px;
+  color: white;
+
+  .buttonContainer {
+    bottom: 10px;
+    position: relative;
+  }
 }
