+++ apps/site-landing/src/containers/BenefitsSimulator/Form/types.ts
@@ -1,0 +1,9 @@
+import { Dispatch, SetStateAction } from 'react';
+import { SimulationParameters } from '../../ReturnProfile/data/params';
+
+export type Props = {
+  params: SimulationParameters;
+  setParams: Dispatch<SetStateAction<SimulationParameters>>;
+  years: number,
+  setYears: Dispatch<SetStateAction<number>>;
+};
